test to calculate IMC of an adult person with no pathology like achondroplasia

+++ src/test/java/dev/erica/imc/imc_calculator/IMCCalculatorTest.java
@@ -2,6 +2,8 @@ package dev.erica.imc.imc_calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.closeTo;
 
 public class IMCCalculatorTest {
     @Test
@@ -10,6 +12,6 @@ public class IMCCalculatorTest {
         double weight = 50;
         double height = 1.60;
 
-        assertThat(IMCCalculator.calculateIMC(weight, height), closeTo(19.53, 0.01));
+        assertThat((Double) IMCCalculator.calculateIMC(weight, height), closeTo(19.53, 0.01));
     }
 }
